move out legacy migration relevant types
